feat(credit): games with top scorer based on their level
This API returns the games along with their gamers who were awarded
maximum credits based on their individual interest levels.
Since a gamer can be awarded a credit for a game s/he had shown
no interest for, the awarded credit will be considered under
an 'UNKNOWN' level. The games for which no gamer is awarded a credit
will not be shown either.